Redefined confirmation form as modal invoked via AJAX. Updated across controllers. Closes #51.

+++ module/Application/src/Controller/JobController.php
@@ -7,6 +7,7 @@ use Zend\Form\Annotation\AnnotationBuilder;
 use Doctrine\ORM\EntityManager;
 use DoctrineModule\Stdlib\Hydrator\DoctrineObject as DoctrineHydrator;
 use Application\Entity\Job;
+use Application\Form\ConfirmationForm;
 
 class JobController extends AbstractActionController
 {
@@ -67,16 +68,44 @@ class JobController extends AbstractActionController
         ));
     }
     
-    public function deleteAction($id)
+    public function deleteAction()
     {   
         $id = (int) $this->params()->fromRoute('id', 0);
         if (!$id) {
             return $this->redirect()->toRoute('jobs');
         }
+
         $job = $this->em->getRepository(Job::class)->find($id);
-        $this->em->remove($job);
-        $this->em->flush();
-        return $this->redirect()->toRoute('jobs');
+        if (!$job) {
+            return $this->redirect()->toRoute('jobs');
+        }
+
+        $builder = new AnnotationBuilder();
+        $form = $builder->createForm(new ConfirmationForm());
+        $form->setAttribute('action', $this->url()->fromRoute('jobs', array('action' => 'delete', 'id' => $id)));
+        $form->get('cancelTo')->setValue($this->url()->fromRoute('jobs'));
+        
+        $request = $this->getRequest();
+        if ($request->isPost()){
+            $form->setData($request->getPost());
+            if ($form->isValid()) { 
+                $data = $form->getData();
+                if ($data['confirm'] == 1) {
+                    $this->em->remove($job);
+                    $this->em->flush();                    
+                } 
+            }
+            return $this->redirect()->toRoute('jobs');
+        } 
+
+        $viewModel = new ViewModel(array(
+            'form' => $form,
+            'entityType' => 'job',
+            'entityDescriptor' => $job->getTitle(),            
+        ));
+        $viewModel->setTerminal($request->isXmlHttpRequest());
+        $viewModel->setTemplate('application/common/confirm.phtml');
+        return $viewModel;
     }
     
 }
