+++ module/Application/src/Controller/LabelController.php
@@ -7,6 +7,7 @@ use Zend\Form\Annotation\AnnotationBuilder;
 use Doctrine\ORM\EntityManager;
 use DoctrineModule\Stdlib\Hydrator\DoctrineObject as DoctrineHydrator;
 use Application\Entity\Label;
+use Application\Form\ConfirmationForm;
 
 class LabelController extends AbstractActionController
 {
@@ -57,16 +58,43 @@ class LabelController extends AbstractActionController
         ));
     }
     
-    public function deleteAction($id)
+    public function deleteAction()
     {   
         $id = (int) $this->params()->fromRoute('id', 0);
         if (!$id) {
             return $this->redirect()->toRoute('labels');
         }
         $label = $this->em->getRepository(Label::class)->find($id);
-        $this->em->remove($label);
-        $this->em->flush();
-        return $this->redirect()->toRoute('labels');
+        if (!$label) {
+            return $this->redirect()->toRoute('labels');
+        } 
+
+        $builder = new AnnotationBuilder();
+        $form = $builder->createForm(new ConfirmationForm());
+        $form->setAttribute('action', $this->url()->fromRoute('labels', array('action' => 'delete', 'id' => $id)));
+        $form->get('cancelTo')->setValue($this->url()->fromRoute('labels'));
+        
+        $request = $this->getRequest();
+        if ($request->isPost()){
+            $form->setData($request->getPost());
+            if ($form->isValid()) { 
+                $data = $form->getData();
+                if ($data['confirm'] == 1) {
+                    $this->em->remove($label);
+                    $this->em->flush();                    
+                } 
+            }
+            return $this->redirect()->toRoute('labels');
+        } 
+
+        $viewModel = new ViewModel(array(
+            'form' => $form,
+            'entityType' => 'label',
+            'entityDescriptor' => $label->getName(),            
+        ));
+        $viewModel->setTerminal($request->isXmlHttpRequest());
+        $viewModel->setTemplate('application/common/confirm.phtml');
+        return $viewModel;
     }
     
 }
